Refactor project structure to follow clean architecture principles
Refactor the project structure to better align with clean architecture principles.

* **Move Service Classes:**
  - Move `BaseApiClientService` and `DashboardApiClientService` from `application/usecases` to `infraestructure/services`.

* **Update Package Declarations and Imports:**
  - Update package declaration and import statements in `GetJiraIssueUseCase` to reflect the new package structure.
  - Update package declaration and import statements in `IJiraIssueRepository` to reflect the new package structure.
  - Update package declaration and import statements in `JiraIssueRepositoryImpl` to reflect the new package structure.

* **Move DTO Classes:**
  - Move `ActorDTO`, `IssueDTO`, `StatusDTO`, `VRComponentDTO`, `VersionDTO`, `ProjectDTO`, and `WorklogDTO` from `domain/entities/jira` to `dto/jira`.

* **Update Main Class Imports:**
  - Update import statements in `Main.java` to reflect the new package structure for `DashboardApiClientService` and `SecureCredentialsLoaderService`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/LoriaLawrenceZ/VRFiscalDashboard?shareId=XXXX-XXXX-XXXX-XXXX).
