+++ src/main/java/br/com/vrsoftware/dto/jira/ActorDTO.java
