+++ src/main/java/br/com/vrsoftware/dto/jira/issue/VRComponentDTO.java
