+++ src/main/java/br/com/vrsoftware/infraestructure/services/BaseApiClientService.java
