+++ src/main/java/br/com/vrsoftware/infraestructure/services/DashboardApiClientService.java
